# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Financial Coach API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vault/accounts:
    get:
      operationId: getVaultAccounts
      tags: [vault]
      summary: Get all accounts in the financial vault
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultAccount"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createVaultAccount
      tags: [vault]
      summary: Create a new account in the vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultAccountInput"
      responses:
        "201":
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultAccount"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /vault/accounts/{id}:
    delete:
      operationId: deleteVaultAccount
      tags: [vault]
      summary: Delete an account from the vault
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Account deleted successfully
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /vault/investments:
    get:
      operationId: getVaultInvestments
      tags: [vault]
      summary: Get all investments in the financial vault
      responses:
        "200":
          description: List of investments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultInvestment"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createVaultInvestment
      tags: [vault]
      summary: Create a new investment in the vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultInvestmentInput"
      responses:
        "201":
          description: Investment created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultInvestment"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /vault/investments/{id}:
    delete:
      operationId: deleteVaultInvestment
      tags: [vault]
      summary: Delete an investment from the vault
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Investment deleted successfully
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Investment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /vault/insurance:
    get:
      operationId: getVaultInsurancePolicies
      tags: [vault]
      summary: Get all insurance policies in the financial vault
      responses:
        "200":
          description: List of insurance policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultInsurancePolicy"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createVaultInsurancePolicy
      tags: [vault]
      summary: Create a new insurance policy in the vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultInsurancePolicyInput"
      responses:
        "201":
          description: Insurance policy created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultInsurancePolicy"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /vault/insurance/{id}:
    delete:
      operationId: deleteVaultInsurancePolicy
      tags: [vault]
      summary: Delete an insurance policy from the vault
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Insurance policy deleted successfully
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Insurance policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    VaultAccount:
      type: object
      properties:
        id:
          type: integer
        institutionName:
          type: string
        accountType:
          type: string
        encryptedAccountMask:
          type: string
        currentBalance:
          type: number
        interestRate:
          type: number
      required: [id, institutionName, accountType, encryptedAccountMask, currentBalance, interestRate]
    VaultAccountInput:
      type: object
      properties:
        institutionName:
          type: string
        accountType:
          type: string
        encryptedAccountMask:
          type: string
        currentBalance:
          type: number
        interestRate:
          type: number
      required: [institutionName, accountType, encryptedAccountMask, currentBalance, interestRate]
    VaultInvestment:
      type: object
      properties:
        id:
          type: integer
        assetClass:
          type: string
        tickerSymbol:
          type: ["string", "null"]
        currentValue:
          type: number
        costBasis:
          type: number
      required: [id, assetClass, currentValue, costBasis]
    VaultInvestmentInput:
      type: object
      properties:
        assetClass:
          type: string
        tickerSymbol:
          type: ["string", "null"]
        currentValue:
          type: number
        costBasis:
          type: number
      required: [assetClass, currentValue, costBasis]
    VaultInsurancePolicy:
      type: object
      properties:
        id:
          type: integer
        policyType:
          type: string
        coverageAmount:
          type: number
        monthlyPremium:
          type: number
      required: [id, policyType, coverageAmount, monthlyPremium]
    VaultInsurancePolicyInput:
      type: object
      properties:
        policyType:
          type: string
        coverageAmount:
          type: number
        monthlyPremium:
          type: number
      required: [policyType, coverageAmount, monthlyPremium]
...